Renamed InputComponent.vue -> TextInput.vue - Moved into @/components/auth/input
