+++ src/views/auth/RegisterPage.vue
@@ -22,7 +22,7 @@
   import { useFetchAPI } from "@/composables/useFetchAPI";
 
   import LogoComponent from "@/components/auth/LogoComponent.vue";
-  import InputComponent from "@/components/auth/InputComponent.vue";
+  import InputComponent from "@/components/auth/input/TextInput.vue";
   import FetchError from "@/utils/errors/FetchError";
 
   const router = useRouter();
